move methods of TestContext related to the login module to a separate file
